// src/globals/events-enum.ts
/**
 * Enumerators for naming common events
 */
export enum myEventsList {
 startTimer = "startTimer",
}

export enum menuActionEvents {
  showHome = "showHome",
  showAbout = "showAbout",
  showMyTableData = "showMyTableData",
}

export enum bodySections {
  home = "home",
  about = "about",
  myTableData = "myTableData",
}